backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
huggingface_hub
qdrant-client
PyYAML